+++ docs/imagecustomizer/api/configuration/verity.md
@@ -259,21 +259,21 @@ Specifies the path where the signed verity hash file should be injected into the
 image. This path is typically used by the `systemd-veritysetup` module to verify
 the verity hash against a signature at boot time.
 
-- This path **must be located under the boot partition** for the current
-  version. For example, if the boot partition is mounted at `/boot`, then
-  `hashSignaturePath: /boot/root.hash.sig` will result in a destination of
-  `/root.hash.sig` relative to the boot partition during injection.
+This path **must be located under the boot partition**. (This restriction may be
+lessened in the future.) For example, if the boot partition is mounted at
+`/boot`, then `hashSignaturePath: /boot/root.hash.sig` will result in a
+destination of `/root.hash.sig` relative to the boot partition during injection.
 
-- When this field is specified, Prism will output the corresponding unsigned
-  hash file (`verity-hash`) as an artifact if the
-  [`output.artifacts`](./outputArtifacts.md) API is configured.
+When this field is specified, Image Customizer will output the corresponding unsigned hash
+file (`verity-hash`) as an artifact if the
+[`output.artifacts`](./outputArtifacts.md) API is configured.
 
-- The generated `inject-files.yaml` will include an entry to inject the signed
-  hash file to the specified path inside the boot partition.
+The generated `inject-files.yaml` will include an entry to inject the signed
+hash file to the specified path inside the boot partition.
 
 If `hashSignaturePath` is not configured for a given Verity entry, the verity
 hash file will not be output even if `verity-hash` is listed in the
-`output.artifacts.items`. Only Verity entries with `hashSignaturePath` defined
-will produce a `verity-hash` artifact.
+[`output.artifacts.items`](./outputArtifacts.md#items-string). Only Verity
+entries with `hashSignaturePath` defined will produce a `verity-hash` artifact.
 
 Added in v0.16.
